@@ -2,83 +2,99 @@
     "name": "knowledge-graph",
     "version": "0.1.0",
     "private": true,
-    "homepage": "/orkg",
+    "homepage": "https://gitlab.com/TIBHannover/orkg/orkg-frontend",
     "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.15",
-        "@fortawesome/free-regular-svg-icons": "^5.8.1",
-        "@fortawesome/free-solid-svg-icons": "^5.7.2",
+        "@fortawesome/fontawesome-svg-core": "^1.2.24",
+        "@fortawesome/free-regular-svg-icons": "^5.11.1",
+        "@fortawesome/free-solid-svg-icons": "^5.11.1",
         "@fortawesome/react-fontawesome": "^0.1.4",
         "array-move": "^2.1.0",
         "babel-preset-es2018": "^1.0.0",
         "body-scroll-lock": "^2.6.4",
-        "bootstrap": "^4.1.3",
+        "bootstrap": "^4.3.1",
         "capitalize": "^2.0.0",
-        "citation-js": "^0.4.8",
+        "citation-js": "^0.4.10",
         "classnames": "^2.2.6",
-        "connected-react-router": "^6.3.2",
-        "css-modules": "^0.3.0",
-        "detect-browser": "^4.6.0",
+        "connected-react-router": "^6.5.2",
+        "detect-browser": "^4.7.0",
         "dot-prop-immutable": "^1.5.0",
         "flow-bin": "^0.95.1",
-        "html2canvas": "^1.0.0-alpha.12",
+        "html2canvas": "1.0.0-alpha.12",
         "jspdf": "^1.5.3",
         "make-latex": "^0.1.0",
         "moment": "^2.24.0",
-        "node-sass": "^4.11.0",
+        "node-sass": "^4.12.0",
         "olap-cube": "^1.0.0",
-        "query-string": "^6.7.0",
+        "query-string": "^6.8.3",
+        "randomcolor": "^0.5.4",
         "rangy": "^1.3.0",
-        "react": "^16.8.6",
+        "react": "^16.9.0",
         "react-app-polyfill": "^1.0.2",
-        "react-content-loader": "^4.2.1",
-        "react-contenteditable": "^3.2.6",
-        "react-cookie": "^4.0.0",
+        "react-content-loader": "^4.2.2",
+        "react-contenteditable": "^3.3.2",
+        "react-cookie": "^4.0.1",
         "react-copy-to-clipboard": "^5.0.1",
         "react-csv": "^1.1.1",
-        "react-dom": "^16.6.0",
-        "react-graph-vis": "^1.0.2",
-        "react-hot-loader": "^4.8.2",
+        "react-dom": "^16.9.0",
+        "react-graph-vis": "^1.0.5",
+        "react-hot-loader": "^4.12.12",
         "react-input-autosize": "^2.2.1",
-        "react-notifications": "^1.4.3",
-        "react-range": "^1.2.0",
+        "react-mathjax-preview-svg": "^1.0.5",
+        "react-number-format": "^4.3.0",
+        "react-range": "^1.3.0",
         "react-redux": "^6.0.1",
         "react-router-config": "^5.0.1",
-        "react-router-dom": "^4.3.1",
-        "react-scripts": "^2.1.8",
+        "react-router-dom": "^5.0.1",
+        "react-scripts": "3.1.2",
         "react-select": "^3.0.4",
-        "react-sortable-hoc": "^1.9.1",
-        "react-split-pane": "^0.1.84",
-        "react-table": "^6.10.0",
+        "react-sortable-hoc": "^1.10.1",
+        "react-split-pane": "^0.1.87",
+        "react-table": "^6.10.3",
         "react-textarea-autosize": "^7.1.0",
-        "react-transition-group": "^1.2.1",
+        "react-toastify": "^5.3.2",
+        "react-transition-group": "^2.9.0",
         "reactour": "^1.15.0",
-        "reactstrap": "^7.1.0",
-        "reactstrap-confirm": "^1.0.4",
-        "redux": "^4.0.1",
+        "reactstrap": "^8.0.1",
+        "reactstrap-confirm": "git://github.com/algm/reactstrap-confirm.git#pull/11/head",
+        "redux": "^4.0.4",
         "redux-thunk": "^2.3.0",
         "strict-mode": "^1.1.3",
-        "styled-components": "^4.1.3",
-        "validator": "^10.11.0",
-        "vis-react": "^0.4.1"
+        "styled-components": "^4.3.2",
+        "validator": "^10.11.0"
     },
     "scripts": {
         "start": "react-scripts start",
         "build": "react-scripts build",
         "test": "react-scripts test --env=jsdom",
         "eject": "react-scripts eject",
-        "flow": "flow"
+        "flow": "flow",
+        "lint": "eslint 'src/**/*.{js,jsx}'",
+        "lint:fix": "eslint 'src/**/*.{js,jsx}' --fix"
     },
     "devDependencies": {
         "babel-preset-env": "^1.7.0",
-        "enzyme": "^3.9.0",
-        "enzyme-adapter-react-16": "^1.12.1",
-        "enzyme-to-json": "^3.3.5",
-        "jest-canvas-mock": "^2.1.0",
+        "enzyme": "^3.10.0",
+        "enzyme-adapter-react-16": "^1.14.0",
+        "enzyme-to-json": "^3.4.0",
+        "husky": "^3.0.7",
+        "jest-canvas-mock": "^2.1.1",
+        "lint-staged": "^9.4.1",
         "named-urls": "^1.4.0",
-        "react-test-renderer": "^16.8.6",
+        "react-test-renderer": "^16.9.0",
         "sass-extract": "^2.1.0",
         "sass-extract-js": "^0.4.0",
         "sass-extract-loader": "^1.1.0"
+    },
+    "lint-staged": {
+        "src/**/*.{js,jsx}": [
+            "npm run --silent lint:fix",
+            "git add"
+        ]
+    },
+    "husky": {
+        "hooks": {
+            "pre-commit": "lint-staged"
+        }
     },
     "browserslist": [
         ">0.2%",
