@@ -1,7 +1,9 @@
 import * as network from '../network';
 import * as type from './types.js';
 import { guid } from '../utils';
+import { mergeWith, isArray } from 'lodash';
 import { createResource, selectResource, createProperty, createValue } from './statementBrowser';
+import { toast } from 'react-toastify';
 
 export const updateGeneralData = (data) => (dispatch) => {
   dispatch({
@@ -19,6 +21,12 @@ export const nextStep = () => (dispatch) => {
 export const previousStep = () => (dispatch) => {
   dispatch({
     type: type.ADD_PAPER_PREVIOUS_STEP,
+  });
+};
+
+export const blockNavigation = () => (dispatch) => {
+  dispatch({
+    type: type.ADD_PAPER_BLOCK_NAVIGATION,
   });
 };
 
@@ -187,15 +195,15 @@ export const prefillStatements = ({ statements, resourceId }) => (dispatch) => {
     }*/
 };
 
-export const deleteContribution = (id) => (dispatch) => {
+export const deleteContribution = (data) => (dispatch) => {
   dispatch({
     type: type.DELETE_CONTRIBUTION,
     payload: {
-      id,
+      id: data.id,
     },
   });
 
-  dispatch(selectContribution({ id: null }));
+  dispatch(selectContribution(data.selectAfterDeletion));
 };
 
 export const selectContribution = (data) => (dispatch) => {
@@ -234,168 +242,104 @@ export const updateResearchProblems = (data) => (dispatch) => {
   });
 };
 
-// Middleware function to transform frontend data to backend format
-// TODO: in order to minimize the amount of requests to the backend, it would be better to make this one backend
-// call that is replaces all the individual requests (backend change is needed for this)
-export const saveAddPaper = (data) => {
-  return async (dispatch) => {
-    const doiPredicate = process.env.REACT_APP_PREDICATES_HAS_DOI;
-    const authorPredicate = process.env.REACT_APP_PREDICATES_HAS_AUTHOR;
-    const publicationMonthPredicate = process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_MONTH;
-    const publicationYearPredicate = process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_YEAR;
-    const researchFieldPredicate = process.env.REACT_APP_PREDICATES_HAS_RESEARCH_FIELD;
-    const contributionPredicate = process.env.REACT_APP_PREDICATES_HAS_CONTRIBUTION;
-    const researchProblemPredicate = process.env.REACT_APP_PREDICATES_HAS_RESEARCH_PROBLEM;
-    const isAPredicate = process.env.REACT_APP_PREDICATES_IS_A;
-    const paperResource = process.env.REACT_APP_RESOURCE_TYPES_PAPER;
+// The function to customize merging objects (to handle using the same existing predicate twice in the same ressource)
+function customizer(objValue, srcValue) {
+  if (isArray(objValue)) {
+    return objValue.concat(srcValue);
+  }
+}
 
-    // title
-    let paper = await network.createResource(data.title);
-
-    // set resource type to paper
-    await network.createResourceStatement(paper.id, isAPredicate, paperResource);
-
-    // DOI
-    let doi = await network.createLiteral(data.doi);
-    await network.createLiteralStatement(paper.id, doiPredicate, doi.id);
-
-    // authors
-    for (let author of data.authors) {
-      // check if the author is already a resource
-      if (author.hasOwnProperty('_class') && author._class === 'resource') {
-        await network.createResourceStatement(paper.id, authorPredicate, author.id);
-      } else {
-        if (author.label) {
-          let authorResource = await network.createResource(author.label);
-          await network.createResourceStatement(paper.id, authorPredicate, authorResource.id);
-        }
-      }
-    }
-
-    if (data.publicationMonth) {
-      // publication month
-      let publicationMonth = await network.createResource(data.publicationMonth);
-      await network.createResourceStatement(paper.id, publicationMonthPredicate, publicationMonth.id);
-    }
-
-    if (data.publicationYear) {
-      // publication year
-      let publicationYear = await network.createResource(data.publicationYear);
-      await network.createResourceStatement(paper.id, publicationYearPredicate, publicationYear.id);
-    }
-
-    // research field
-    await network.createResourceStatement(paper.id, researchFieldPredicate, data.selectedResearchField);
-
-    let contributionIds = [];
-    // contributions
-    for (let contributionId of data.contributions.allIds) {
-      let contribution = data.contributions.byId[contributionId];
-      let contributionResource = await network.createResource(contribution.label);
-      await network.createResourceStatement(paper.id, contributionPredicate, contributionResource.id);
-
-      contributionIds.push(contributionResource.id);
-
-      // set the id of the just created contribution for the related resource
-      data.resources.byId[contribution.resourceId].existingResourceId = contributionResource.id;
-      data.resources.byId[contribution.resourceId].partOfContribution = true;
-
-      // research problems
-      if (contribution.researchProblems && contribution.researchProblems.length > 0) {
-        for (let researchProblem of contribution.researchProblems) {
-          // check if the research problem is already a resource
-          if (researchProblem.hasOwnProperty('_class') && researchProblem._class === 'resource') {
-            await network.createResourceStatement(contributionResource.id, researchProblemPredicate, researchProblem.id);
-          } else {
-            let researchProblemResource = await network.createResource(researchProblem.id);
-            await network.createResourceStatement(contributionResource.id, researchProblemPredicate, researchProblemResource.id);
+export const getResourceObject = (data, resourceId) => {
+  // Make a list of new resources ids
+  let newResources = data.values.allIds.filter(valueId => !data.values.byId[valueId].isExistingValue).map(valueId => data.values.byId[valueId].resourceId);
+  return mergeWith({}, ...data.resources.byId[resourceId].propertyIds.map(propertyId => {
+    let property = data.properties.byId[propertyId];
+    return {
+      // Map properties of resource
+      [property.existingPredicateId ? property.existingPredicateId : `_${propertyId}`]: property.valueIds.map((valueId => {
+        let value = data.values.byId[valueId];
+        if (value.type === 'literal' && !value.isExistingValue) {
+          return {
+            'text': value.label,
           }
-        }
-      }
-    }
-
-    await saveStatements(data);
-
-    for (let contributionId of contributionIds) {
-      // index contribution for similarity
-      network.indexContribution(contributionId);
-    }
-
-    dispatch({
-      type: type.SAVE_ADD_PAPER,
-      id: paper.id,
-    });
-  };
-};
-
-// Maybe this function needs to be resursive
-export const saveStatements = async (data) => {
-  let newProperties = {};
-  let newResources = {};
-
-  if (data.resources.byId) {
-    for (let resourceIdArray of data.resources.allIds) { // the order matters here, since it is an hierarch, the object doesn't provide a reliable order (so use the array)
-      let resource = data.resources.byId[resourceIdArray];
-      let resourceId;
-
-      if (!resource.existingResourceId) {
-        resourceId = await network.createResource(resource.label);
-        resourceId = resourceId.id;
-        data.resources.byId[resourceIdArray].existingResourceId = resourceId;
-      } else {
-        resourceId = resource.existingResourceId;
-      }
-
-      // predicates
-      if (resource.propertyIds && resource.propertyIds.length > 0) {
-        for (let propertyId of resource.propertyIds) {
-          let property = data.properties.byId[propertyId];
-
-          let predicateId;
-
-          if (!property.existingPredicateId && !newProperties[property.label]) {
-            predicateId = await network.createPredicate(property.label);
-            predicateId = predicateId.id;
-
-            newProperties[property.label] = predicateId; // add to the newProperties object, so the ID can be used for other new labels
-          } else if (newProperties[property.label]) {
-            predicateId = newProperties[property.label];
+        } else {
+          if (!value.isExistingValue) {
+            let newResources = {}
+            newResources[value.resourceId] = value.resourceId;
+            return {
+              '@temp': `_${value.resourceId}`,
+              'label': value.label,
+              'values': Object.assign({}, getResourceObject(data, value.resourceId))
+            }
           } else {
-            predicateId = property.existingPredicateId;
-          }
-
-          // objects/values
-          if (property.valueIds && property.valueIds.length > 0) {
-            for (let valueId of property.valueIds) {
-              let value = data.values.byId[valueId];
-
-              if (value.type === 'literal' && !value.isExistingValue) {
-                let newValueId = await network.createLiteral(value.label);
-                newValueId = newValueId.id;
-
-                await network.createLiteralStatement(resourceId, predicateId, newValueId);
-              } else {
-                let newValueId;
-
-                if (!value.isExistingValue) {
-                  newValueId = await network.createResource(value.label);
-                  newValueId = newValueId.id;
-                  data.resources.byId[value.resourceId].existingResourceId = newValueId;
-                  newResources[value.resourceId] = newValueId;
-                } else if (newResources[value.resourceId]) {
-                  newValueId = newResources[value.resourceId];
-                } else {
-                  newValueId = value.resourceId;
-                }
-
-                if (!value.existingStatement) {
-                  await network.createResourceStatement(resourceId, predicateId, newValueId);
-                }
-              }
+            return {
+              '@id': newResources.includes(value.resourceId) ? `_${value.resourceId}` : value.resourceId,
             }
           }
         }
+      }))
+    }
+  }), customizer)
+}
+
+// Middleware function to transform frontend data to backend format
+export const saveAddPaper = (data) => {
+  return async (dispatch) => {
+    const researchProblemPredicate = process.env.REACT_APP_PREDICATES_HAS_RESEARCH_PROBLEM;
+
+    let paperObj = {
+      // Set new predicates label and temp ID
+      'predicates': data.properties.allIds.filter(propertyId => !data.properties.byId[propertyId].existingPredicateId).map(propertyId => {
+        let property = data.properties.byId[propertyId];
+        return {
+          [property.label]: `_${propertyId}`
+        }
+      }),
+      // Set the paper metadata
+      'paper': {
+        'title': data.title,
+        'doi': data.doi,
+        'authors': data.authors.map(author => ({ label: author.label })),
+        'publicationMonth': data.publicationMonth,
+        'publicationYear': data.publicationYear,
+        'researchField': data.selectedResearchField,
+        // Set the contributions data
+        'contributions': data.contributions.allIds.map(c => {
+          let contribution = data.contributions.byId[c];
+          let researhProblem = {
+            [researchProblemPredicate]: contribution.researchProblems.map(rp => {
+              if (rp.hasOwnProperty('_class') && rp._class === 'resource') {
+                return { '@id': rp.id }
+              } else {
+                return { label: rp.label }
+              }
+            })
+          }
+          return {
+            'name': contribution.label,
+            'values': Object.assign({}, researhProblem, getResourceObject(data, contribution.resourceId))
+          }
+        })
       }
     }
-  }
+
+    try {
+      let paper = await network.saveFullPaper(paperObj);
+      dispatch({
+        type: type.SAVE_ADD_PAPER,
+        id: paper.id,
+      });
+
+      dispatch({
+        type: type.ADD_PAPER_UNBLOCK_NAVIGATION
+      });
+    } catch (e) {
+      console.log(e);
+      toast.error('Something went wrong while saving this paper.');
+      dispatch({
+        type: type.ADD_PAPER_PREVIOUS_STEP
+      });
+    }
+
+  };
 };
