@@ -14,7 +14,7 @@ import {
 } from '../../../actions/addPaper';
 import { withCookies, Cookies } from 'react-cookie';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
-import { faSpinner } from '@fortawesome/free-solid-svg-icons';
+import { faSpinner, faQuestionCircle } from '@fortawesome/free-solid-svg-icons';
 import AbstractAnnotator from './AbstractAnnotator';
 import PropTypes from 'prop-types';
 import Textarea from 'react-textarea-autosize';
@@ -26,6 +26,9 @@ import { Range, getTrackBackground } from 'react-range';
 import Tour from 'reactour';
 import toArray from 'lodash/toArray';
 import { disableBodyScroll, enableBodyScroll, clearAllBodyScrollLocks } from 'body-scroll-lock';
+import randomcolor from 'randomcolor';
+import capitalize from 'capitalize';
+
 
 class Abstract extends Component {
   constructor(props) {
@@ -40,9 +43,22 @@ class Abstract extends Component {
       showError: false,
       changeAbstract: false,
       classOptions: [],
-      certaintyThreshold: [0.8],
+      classColors: {
+        'process': '#7fa2ff',
+        'data': '	#9df28a',
+        'material': '#EAB0A2',
+        'method': '#D2B8E5',
+      },
+      certaintyThreshold: [0.5],
       validation: true,
     };
+
+    this.automaticAnnotationConcepts = [
+      { label: 'Process', description: 'Natural phenomenon, or independent/dependent activities.E.g., growing(Bio), cured(MS), flooding(ES).' },
+      { label: 'Data', description: 'The data themselves, or quantitative or qualitative characteristics of entities. E.g., rotational energy (Eng), tensile strength (MS), the Chern character (Mat).' },
+      { label: 'Material', description: 'A physical or digital entity used for scientific experiments. E.g., soil (Agr), the moon (Ast), the set (Mat).' },
+      { label: 'Method', description: 'A commonly used procedure that acts on entities. E.g., powder X-ray (Che), the PRAM analysis (CS), magnetoencephalography (Med).' }
+    ];
 
     // check if a cookie of take a tour exist 
     if (this.props.cookies && this.props.cookies.get('taketour') === 'take' && this.props.tourCurrentStep === 1
@@ -66,13 +82,12 @@ class Abstract extends Component {
 
   loadClassOptions = () => {
     // Fetch the predicates used in the NLP model
-    let classeOptions = ['Process', 'Data', 'Material', 'Method'];
-    let nLPPredicates = classeOptions.map((classOption) => {
-      return submitGetRequest(predicatesUrl + '?q=' + classOption + '&exact=true').then(predicates => {
+    let nLPPredicates = this.automaticAnnotationConcepts.map((classOption) => {
+      return submitGetRequest(predicatesUrl + '?q=' + classOption.label + '&exact=true').then(predicates => {
         if (predicates.length > 0) {
           return predicates[0]; // Use the first predicate that match the label
         } else {
-          return createPredicate(classOption) // Create the predicate if it doesn't exist
+          return createPredicate(classOption.label) // Create the predicate if it doesn't exist
         }
       })
     })
@@ -144,7 +159,12 @@ class Abstract extends Component {
 
   fetchAbstract = async () => {
     if (!this.props.abstract) {
-      let DOI = this.props.doi.substring(this.props.doi.indexOf('10.'));
+      let DOI;
+      try {
+        DOI = this.props.doi.substring(this.props.doi.indexOf('10.'));
+      } catch{
+        DOI = false
+      }
       if (!this.props.title || !DOI) {
         this.setState({
           changeAbstract: true,
@@ -294,6 +314,19 @@ class Abstract extends Component {
     }
   };
 
+  getClassColor = (rangeClass) => {
+    if (!rangeClass) {
+      return '#ffb7b7';
+    }
+    if (this.state.classColors[rangeClass.toLowerCase()]) {
+      return this.state.classColors[rangeClass.toLowerCase()];
+    } else {
+      let newColor = randomcolor({ luminosity: 'light', seed: rangeClass.toLowerCase() });
+      this.setState({ classColors: { ...this.state.classColors, [rangeClass.toLowerCase()]: newColor } });
+      return newColor;
+    }
+  }
+
   render() {
     let rangeArray = toArray(this.props.ranges).filter(
       (r) => (r.certainty >= this.state.certaintyThreshold)
@@ -361,38 +394,36 @@ class Abstract extends Component {
                       <span className={'mr-1 ml-1'} />
                       {rangesClasses.length > 0 &&
                         rangesClasses.map((c) => {
-                          let color = '#0052CC';
-                          switch (c) {
-                            case 'Process':
-                              color = '#7fa2ff';
-                              break;
-                            case 'Data':
-                              color = '#5FA97F';
-                              break;
-                            case 'Material':
-                              color = '#EAB0A2';
-                              break;
-                            case 'Method':
-                              color = '#D2B8E5';
-                              break;
-                            default:
-                              color = '#ffb7b7';
+                          let aconcept = c ? this.automaticAnnotationConcepts.filter(function (e) { return e.label.toLowerCase() === c.toLowerCase(); }) : []
+                          if (c && aconcept.length > 0) {
+                            return (
+                              <Tooltip hideDefaultIcon={true} message={aconcept[0].description}>
+                                <Badge
+                                  className={'mr-2'}
+                                  key={`c${c}`}
+                                  style={{ cursor: 'pointer', marginBottom: '4px', color: '#333', background: this.getClassColor(c) }}
+                                >
+                                  {c ? capitalize(c) : 'Unlabeled'} <Badge pill color="secondary">{rangeArray.filter((rc) => rc.class.label === c).length}</Badge>
+                                </Badge>
+                              </Tooltip>
+                            );
+                          } else {
+                            return (
+                              <Badge
+                                className={'mr-2'}
+                                key={`c${c}`}
+                                style={{ marginBottom: '4px', color: '#333', background: this.getClassColor(c) }}
+                              >
+                                {c ? capitalize(c) : 'Unlabeled'} <Badge pill color="secondary">{rangeArray.filter((rc) => rc.class.label === c).length}</Badge>
+                              </Badge>
+                            );
                           }
-                          //
-                          return (
-                            <Badge
-                              className={'mr-2'}
-                              key={`c${c}`}
-                              style={{ color: '#333', background: color }}
-                            >
-                              {c ? c : 'Unlabeled'} <Badge pill color="secondary">{rangeArray.filter((rc) => rc.class.label === c).length}</Badge>
-                            </Badge>
-                          );
                         })}
                     </div>
                     <AbstractAnnotator
                       certaintyThreshold={this.state.certaintyThreshold[0]}
                       classOptions={this.state.classOptions}
+                      getClassColor={this.getClassColor}
                     />
                   </div>
                 )}
@@ -499,8 +530,9 @@ class Abstract extends Component {
                   )}
                 />
                 <div className={'mt-2 text-center'}>
-                  <Tooltip message="Here you can adjust the certainty value, that means at which level you accept the confidence ratio of automatic annotations. Only the shown annotations will be used to create the contribution data in the next step.">
-                    Certainty {this.state.certaintyThreshold[0].toFixed(2)}
+                  <span className={'mr-2'}>Certainty {this.state.certaintyThreshold[0].toFixed(2)}</span>
+                  <Tooltip trigger={'click'} hideDefaultIcon={true} message="Here you can adjust the certainty value, that means at which level you accept the confidence ratio of automatic annotations. Only the shown annotations will be used to create the contribution data in the next step.">
+                    <Icon style={{ cursor: 'pointer' }} className={'text-primary'} icon={faQuestionCircle} />
                   </Tooltip>
                 </div>
               </div>
