@@ -12,11 +12,12 @@ import Finish from './Finish/Finish';
 import { withCookies } from 'react-cookie';
 import { connect } from 'react-redux';
 import styled, { withTheme } from 'styled-components';
-import { CSSTransitionGroup } from 'react-transition-group';
+import { TransitionGroup, CSSTransition } from 'react-transition-group';
 import PropTypes from 'prop-types';
 import { resetStatementBrowser } from '../../actions/statementBrowser';
-import { openTour, closeTour } from '../../actions/addPaper';
+import { openTour, closeTour, blockNavigation } from '../../actions/addPaper';
 import GraphViewModal from '../ViewPaper/GraphViewModal';
+import { Prompt } from 'react-router'
 
 const Help = styled.div`
     box-sizing: border-box;
@@ -56,7 +57,7 @@ const HelpIcon = styled(Icon)`
     font-size: 18px;
 `;
 
-const AnimationContainer = styled.div`
+const AnimationContainer = styled(CSSTransition)`
     &.fadeIn-enter {
         opacity: 0;
     }
@@ -65,6 +66,24 @@ const AnimationContainer = styled.div`
         opacity: 1;
         transition: 1s opacity;
     }
+`;
+
+const SubtitleSeperator = styled.div`
+    background: ${(props) => props.theme.darkblue};
+    width:2px;
+    height:30px;
+    margin:0 15px;
+    content: "";
+    display: block;
+    opacity: 0.7;
+`;
+
+const PaperTitle = styled.div`
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    margin-right:20px;
+    color: #62687d;
 `;
 
 class AddPaper extends Component {
@@ -82,6 +101,21 @@ class AddPaper extends Component {
         document.title = 'Add paper - ORKG';
 
         this.props.resetStatementBrowser();
+    }
+
+    componentDidUpdate(prevProps) {
+        //paperNewResourceId : means paper is saved
+        if (!this.props.shouldBlockNavigation && this.props.currentStep > 1 && !this.props.paperNewResourceId) {
+            this.props.blockNavigation();
+            window.onbeforeunload = () => true;
+        }
+        if (!this.props.shouldBlockNavigation && prevProps.shouldBlockNavigation !== this.props.shouldBlockNavigation) {
+            window.onbeforeunload = null;
+        }
+    }
+
+    componentWillUnmount() {
+        window.onbeforeunload = null;
     }
 
     toggleDropdown = () => {
@@ -112,33 +146,47 @@ class AddPaper extends Component {
             case 1:
             default:
                 currentStepDetails = (
-                    <AnimationContainer key={1}>
+                    <AnimationContainer key={1} classNames="fadeIn" timeout={{ enter: 700, exit: 0 }}>
                         <GeneralData />
                     </AnimationContainer>
                 );
                 break;
             case 2:
                 currentStepDetails = (
-                    <AnimationContainer key={2}>
+                    <AnimationContainer key={2} classNames="fadeIn" timeout={{ enter: 700, exit: 0 }}>
                         <ResearchField />
                     </AnimationContainer>
                 );
                 break;
             case 3:
-                currentStepDetails = <AnimationContainer key={3}><Abstract /></AnimationContainer>
+                currentStepDetails = <AnimationContainer key={3} classNames="fadeIn" timeout={{ enter: 700, exit: 0 }}><Abstract /></AnimationContainer>
                 break;
             case 4:
-                currentStepDetails = <AnimationContainer key={4}><Contributions /></AnimationContainer>
+                currentStepDetails = <AnimationContainer key={4} classNames="fadeIn" timeout={{ enter: 700, exit: 0 }}><Contributions /></AnimationContainer>
                 break;
             case 5:
-                currentStepDetails = <AnimationContainer key={5}><Finish /></AnimationContainer>
+                currentStepDetails = <AnimationContainer key={5} classNames="fadeIn" timeout={{ enter: 700, exit: 0 }}><Finish /></AnimationContainer>
                 break;
         }
 
         return (
             <div>
-                <Container className="p-0">
-                    <h1 className="h4 mt-4 mb-4 float-left">Add paper</h1>
+                <Prompt
+                    when={this.props.shouldBlockNavigation}
+                    message="You have unsaved changes, are you sure you want to leave?"
+                />
+                <Container className="p-0 mt-4 mb-4 d-flex align-items-center">
+                    <h1 className="h4 flex-shrink-0 mb-0">Add paper</h1>
+
+                    {this.props.currentStep > 1 && (
+                        <>
+                            <SubtitleSeperator />
+
+                            <PaperTitle>
+                                {this.props.title}
+                            </PaperTitle>
+                        </>
+                    )}
 
                     {/*<Dropdown isOpen={this.state.dropdownOpen} toggle={this.toggleDropdown} className="mb-4 mt-4 float-right" style={{ marginLeft: 'auto' }}>
                         <DropdownToggle color="darkblue" size="sm">
@@ -149,15 +197,15 @@ class AddPaper extends Component {
                         </DropdownMenu>
                     </Dropdown>*/}
 
-                        <Button 
-                            color="darkblue" 
-                            size="sm" 
-                            className="mb-4 mt-4 float-right" 
-                            style={{ marginLeft: 'auto' }} 
-                            onClick={() => this.toggle('showGraphModal')}
-                        >
-                            <Icon icon={faProjectDiagram} className="mr-1" /> View graph
-                        </Button>
+                    <Button
+                        color="darkblue"
+                        size="sm"
+                        className="flex-shrink-0"
+                        style={{ marginLeft: 'auto' }}
+                        onClick={() => this.toggle('showGraphModal')}
+                    >
+                        <Icon icon={faProjectDiagram} className="mr-1" /> View graph
+                    </Button>
 
                     <div className="clearfix" />
                 </Container>
@@ -166,9 +214,9 @@ class AddPaper extends Component {
 
                     <hr />
 
-                    <CSSTransitionGroup transitionName="fadeIn" transitionEnterTimeout={700} transitionLeave={false}>
+                    <TransitionGroup exit={false}>
                         {currentStepDetails}
-                    </CSSTransitionGroup>
+                    </TransitionGroup>
                 </Container>
 
                 <GraphViewModal
@@ -189,29 +237,39 @@ class AddPaper extends Component {
                 </Help>
 
 
-            </div>
+            </div >
         );
     }
 }
 
 AddPaper.propTypes = {
     currentStep: PropTypes.number.isRequired,
+    title: PropTypes.string.isRequired,
+    authors: PropTypes.array.isRequired,
     isTourOpen: PropTypes.bool.isRequired,
+    shouldBlockNavigation: PropTypes.bool.isRequired,
     resetStatementBrowser: PropTypes.func.isRequired,
+    paperNewResourceId: PropTypes.string,
     openTour: PropTypes.func.isRequired,
     closeTour: PropTypes.func.isRequired,
+    blockNavigation: PropTypes.func.isRequired,
     theme: PropTypes.object.isRequired
 };
 
 const mapStateToProps = (state) => ({
     currentStep: state.addPaper.currentStep,
     isTourOpen: state.addPaper.isTourOpen,
+    title: state.addPaper.title,
+    authors: state.addPaper.authors,
+    shouldBlockNavigation: state.addPaper.shouldBlockNavigation,
+    paperNewResourceId: state.addPaper.paperNewResourceId,
 });
 
 const mapDispatchToProps = (dispatch) => ({
     resetStatementBrowser: () => dispatch(resetStatementBrowser()),
     openTour: () => dispatch(openTour()),
     closeTour: () => dispatch(closeTour()),
+    blockNavigation: () => dispatch(blockNavigation()),
 });
 
 export default compose(
