@@ -13,14 +13,14 @@ import {
 } from '../../../actions/addPaper';
 import Confirm from 'reactstrap-confirm';
 import Contribution from './Contribution';
-import { CSSTransitionGroup } from 'react-transition-group'
+import { TransitionGroup, CSSTransition } from 'react-transition-group';
 import styled, { withTheme } from 'styled-components';
 import { withCookies } from 'react-cookie';
 import PropTypes from 'prop-types';
 
 
 
-const AnimationContainer = styled.div`
+const AnimationContainer = styled(CSSTransition)`
     transition: 0.3s background-color,  0.3s border-color;
 
     &.fadeIn-enter {
@@ -96,7 +96,9 @@ class Contributions extends Component {
         });
 
         if (result) {
-            this.props.deleteContribution(id);
+            // delete the contribution and select the first one in the remaining list 
+            let selectedId = this.props.contributions.allIds.filter(i => i !== id)[0];
+            this.props.deleteContribution({ id, selectAfterDeletion: this.props.contributions.byId[selectedId] });
         }
     }
 
@@ -191,26 +193,22 @@ class Contributions extends Component {
                             </StyledContributionsList>
                         </Col>
 
-                        <CSSTransitionGroup
-                            transitionName="fadeIn"
-                            transitionEnterTimeout={500}
-                            transitionLeave={false}
-                            component="div"
-                            className="col-9"
-                        >
+                        <TransitionGroup className="col-9" exit={false}>
                             <AnimationContainer
+                                classNames="fadeIn"
+                                timeout={{ enter: 500, exit: 0 }}
                                 key={selectedResourceId}
                             >
                                 <Contribution id={selectedResourceId} />
                             </AnimationContainer>
-                        </CSSTransitionGroup>
+                        </TransitionGroup>
                     </Row>
                 </Container>
 
                 <hr className="mt-5 mb-3" />
                 <Button color="primary" className="float-right mb-4" onClick={this.handleNextClick}>Finish</Button>
                 <Button color="light" className="float-right mb-4 mr-2" onClick={this.props.previousStep}>Previous step</Button>
-            </div>
+            </div >
         );
     }
 }
@@ -218,8 +216,14 @@ class Contributions extends Component {
 Contributions.propTypes = {
     title: PropTypes.string.isRequired,
     authors: PropTypes.array.isRequired,
-    publicationMonth: PropTypes.number.isRequired,
-    publicationYear: PropTypes.number.isRequired,
+    publicationMonth: PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.number
+    ]).isRequired,
+    publicationYear: PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.number
+    ]).isRequired,
     doi: PropTypes.string.isRequired,
     selectedResearchField: PropTypes.string.isRequired,
     contributions: PropTypes.object.isRequired,
@@ -259,7 +263,7 @@ const mapDispatchToProps = dispatch => ({
     nextStep: () => dispatch(nextStep()),
     previousStep: () => dispatch(previousStep()),
     createContribution: (data) => dispatch(createContribution(data)),
-    deleteContribution: (id) => dispatch(deleteContribution(id)),
+    deleteContribution: (data) => dispatch(deleteContribution(data)),
     selectContribution: (id) => dispatch(selectContribution(id)),
     updateContributionLabel: (data) => dispatch(updateContributionLabel(data)),
     saveAddPaper: (data) => dispatch(saveAddPaper(data)),
