@@ -12,7 +12,7 @@ import AuthorsInput from '../../Utils/AuthorsInput';
 import FormValidator from '../../Utils/FormValidator';
 import { connect } from 'react-redux';
 import { updateGeneralData, nextStep, openTour, closeTour, updateTourCurrentStep } from '../../../actions/addPaper';
-import { CSSTransitionGroup } from 'react-transition-group';
+import { TransitionGroup, CSSTransition } from 'react-transition-group';
 import { withCookies, Cookies } from 'react-cookie';
 import styled, { withTheme } from 'styled-components';
 import moment from 'moment';
@@ -21,7 +21,7 @@ import Cite from 'citation-js';
 import Tour from 'reactour';
 import { disableBodyScroll, enableBodyScroll, clearAllBodyScrollLocks } from 'body-scroll-lock';
 
-const Container = styled.div`
+const Container = styled(CSSTransition)`
   &.fadeIn-enter {
     opacity: 0;
   }
@@ -31,7 +31,7 @@ const Container = styled.div`
     transition: 1s opacity;
   }
 
-  &.fadeIn-leave.fadeIn-leave-active {
+  &.fadeIn-exit.fadeIn-exit-active {
     display: none;
   }
 
@@ -45,7 +45,7 @@ const Container = styled.div`
     transition: 1s;
   }
 
-  &.slideDown-leave.slideDown-leave-active {
+  &.slideDown-exit.slideDown-exit-active {
     display: none;
   }
 `;
@@ -79,6 +79,7 @@ class GeneralData extends Component {
             paperPublicationMonth: this.props.publicationMonth,
             paperPublicationYear: this.props.publicationYear,
             validation: this.validator.valid(),
+            errors: null
         };
 
         // Hide the tour if a cookie 'taketour' exist 
@@ -104,7 +105,7 @@ class GeneralData extends Component {
 
         this.lookup.current.blur();
 
-        let validation = this.validator.validate({ entry: this.state.entry });
+        let validation = this.validator.validate({ entry: this.state.entry.trim() });
         this.setState({ validation });
 
         if (!validation.isValid) {
@@ -115,7 +116,14 @@ class GeneralData extends Component {
             isFetching: true,
         });
 
-        await Cite.async(this.state.entry)
+        let entry;
+        if (this.state.entry.trim().startsWith('http')) {
+            entry = this.state.entry.trim().substring(this.state.entry.trim().indexOf('10.'));
+        } else {
+            entry = this.state.entry.trim();
+        }
+
+        await Cite.async(entry)
             .catch((e) => {
                 let validation;
                 switch (e.message) {
@@ -141,8 +149,8 @@ class GeneralData extends Component {
                 }
                 this.setState({
                     isFetching: false,
-                    paperTitle: this.state.entry, // Probably the user entered a paper title
                     validation,
+                    errors: null
                 });
                 return null;
             })
@@ -150,25 +158,34 @@ class GeneralData extends Component {
                 if (paper) {
                     let paperTitle = '',
                         paperAuthors = [],
-                        paperPublicationMonth = null,
-                        paperPublicationYear = null,
+                        paperPublicationMonth = '',
+                        paperPublicationYear = '',
                         doi = '';
                     try {
                         paperTitle = paper.data[0].title;
-                        paperAuthors = paper.data[0].author.map((author, index) => {
-                            let fullname = [author.given, author.family].join(' ').trim();
-                            if (!fullname) {
-                                fullname = author.literal ? author.literal : '';
-                            }
-                            const newAuthor = {
-                                label: fullname,
-                                id: fullname,
-                            };
-                            return newAuthor;
-                        });
-                        paperPublicationMonth = paper.data[0].issued['date-parts'][0][1];
-                        paperPublicationYear = paper.data[0].issued['date-parts'][0][0];
-                        doi = paper.data[0].DOI;
+                        if (paper.data[0].subtitle && paper.data[0].subtitle.length > 0) { // include the subtitle
+                            paperTitle = `${paperTitle}: ${paper.data[0].subtitle[0]}`
+                        }
+                        if (paper.data[0].author) {
+                            paperAuthors = paper.data[0].author.map((author, index) => {
+                                let fullname = [author.given, author.family].join(' ').trim();
+                                if (!fullname) {
+                                    fullname = author.literal ? author.literal : '';
+                                }
+                                const newAuthor = {
+                                    label: fullname,
+                                    id: fullname,
+                                };
+                                return newAuthor;
+                            });
+                        }
+                        if (paper.data[0].issued['date-parts'][0][1]) {
+                            paperPublicationMonth = paper.data[0].issued['date-parts'][0][1];
+                        }
+                        if (paper.data[0].issued['date-parts'][0][0]) {
+                            paperPublicationYear = paper.data[0].issued['date-parts'][0][0];
+                        }
+                        doi = paper.data[0].DOI ? paper.data[0].DOI : '';
                     } catch (e) {
                         console.log('Error setting paper data: ', e);
                     }
@@ -181,6 +198,7 @@ class GeneralData extends Component {
                         paperPublicationMonth,
                         paperPublicationYear,
                         doi: doi,
+                        errors: null
                     });
                 }
             });
@@ -245,6 +263,8 @@ class GeneralData extends Component {
 
     handleNextClick = () => {
         // TODO do some sort of validation, before proceeding to the next step
+        let errors = [];
+
         let {
             paperTitle,
             paperAuthors,
@@ -255,17 +275,25 @@ class GeneralData extends Component {
             showLookupTable,
         } = this.state;
 
-        this.props.updateGeneralData({
-            title: paperTitle,
-            authors: paperAuthors,
-            publicationMonth: paperPublicationMonth,
-            publicationYear: paperPublicationYear,
-            doi: doi,
-            entry: entry,
-            showLookupTable: showLookupTable
-        });
+        if (!paperTitle || paperTitle.trim().length < 1) {
+            errors.push('Please enter the title of your paper or click on "Lookup" if you entered the doi.');
+        }
 
-        this.props.nextStep();
+        if (errors.length === 0) {
+            this.props.updateGeneralData({
+                title: paperTitle,
+                authors: paperAuthors,
+                publicationMonth: paperPublicationMonth,
+                publicationYear: paperPublicationYear,
+                doi: doi,
+                entry: entry,
+                showLookupTable: showLookupTable
+            });
+            this.props.nextStep();
+        }
+        else {
+            this.setState({ errors: errors })
+        }
     };
 
     submitHandler = (e) => {
@@ -312,119 +340,122 @@ class GeneralData extends Component {
                     </Button>
                 </ButtonGroup>
 
-                <CSSTransitionGroup
-                    transitionName="fadeIn"
-                    transitionEnterTimeout={500}
-                    transitionLeave={false}
-                >
+                <TransitionGroup exit={false}>
                     {(() => {
                         switch (this.state.dataEntry) {
                             case 'doi':
                                 return (
-                                    <Container key={1}>
-                                        <Form className="mt-4" onSubmit={this.submitHandler}>
-                                            <FormGroup>
-                                                <Label for="paperDoi">
-                                                    <Tooltip message={<span>Automatically fetch the details of your paper by providing a DOI or a BibTeX entry. <span style={{ textDecoration: 'underline', cursor: 'pointer' }} onClick={() => this.handleLearnMore(0)}>Learn more</span></span>}>
-                                                        Paper DOI or BibTeX
-                                                    </Tooltip>
-                                                </Label>
-                                                <InputGroup id="doiInputGroup">
-                                                    <Input
-                                                        type="text"
-                                                        name="entry"
-                                                        id="paperDoi"
-                                                        value={this.state.entry}
-                                                        onChange={this.handleInputChange}
-                                                        invalid={this.state.validation.entry.isInvalid}
-                                                        onKeyPress={(target) => { target.charCode === 13 && this.handleLookupClick(); }}
-                                                    />
-                                                    <FormFeedback className="order-1">
-                                                        {this.state.validation.entry.message}
-                                                    </FormFeedback>{' '}
-                                                    {/* Need to set order-1 here to fix Bootstrap bug of missing rounded borders */}
-                                                    <InputGroupAddon addonType="append">
-                                                        <Button
-                                                            outline
-                                                            color="primary"
-                                                            innerRef={this.lookup}
-                                                            style={{ minWidth: 130 }}
-                                                            onClick={this.handleLookupClick}
-                                                            disabled={this.state.isFetching}
-                                                            data-test="lookupDoi"
-                                                        >
-                                                            {!this.state.isFetching ? (
-                                                                'Lookup'
-                                                            ) : (
-                                                                    <FontAwesomeIcon icon={faSpinner} spin />
-                                                                )}
-                                                        </Button>
-                                                    </InputGroupAddon>
-                                                </InputGroup>
-                                            </FormGroup>
-                                        </Form>
+                                    <Container key={1} classNames="fadeIn" timeout={{ enter: 500, exit: 0 }}>
+                                        <div>
+                                            <Form className="mt-4" onSubmit={this.submitHandler}>
+                                                <FormGroup>
+                                                    <Label for="paperDoi">
+                                                        <Tooltip message={<span>Automatically fetch the details of your paper by providing a DOI or a BibTeX entry. <span style={{ textDecoration: 'underline', cursor: 'pointer' }} onClick={() => this.handleLearnMore(0)}>Learn more</span></span>}>
+                                                            Paper DOI or BibTeX
+                                                        </Tooltip>
+                                                    </Label>
+                                                    <InputGroup id="doiInputGroup">
+                                                        <Input
+                                                            type="text"
+                                                            name="entry"
+                                                            id="paperDoi"
+                                                            value={this.state.entry}
+                                                            onChange={this.handleInputChange}
+                                                            invalid={this.state.validation.entry.isInvalid}
+                                                            onKeyPress={(target) => { target.charCode === 13 && this.handleLookupClick(); }}
+                                                        />
+                                                        <FormFeedback className="order-1">
+                                                            {this.state.validation.entry.message}
+                                                        </FormFeedback>{' '}
+                                                        {/* Need to set order-1 here to fix Bootstrap bug of missing rounded borders */}
+                                                        <InputGroupAddon addonType="append">
+                                                            <Button
+                                                                outline
+                                                                color="primary"
+                                                                innerRef={this.lookup}
+                                                                style={{ minWidth: 130 }}
+                                                                onClick={this.handleLookupClick}
+                                                                disabled={this.state.isFetching}
+                                                                data-test="lookupDoi"
+                                                            >
+                                                                {!this.state.isFetching ? (
+                                                                    'Lookup'
+                                                                ) : (
+                                                                        <FontAwesomeIcon icon={faSpinner} spin />
+                                                                    )}
+                                                            </Button>
+                                                        </InputGroupAddon>
+                                                    </InputGroup>
+                                                </FormGroup>
+                                            </Form>
 
-                                        <CSSTransitionGroup
-                                            transitionName="slideDown"
-                                            transitionEnterTimeout={500}
-                                            transitionLeaveTimeout={300}
-                                        >
-                                            {this.state.showLookupTable ? (
-                                                <Container key={1} className="mt-5">
-                                                    <h3 className="h4 mb-3">
-                                                        Lookup result
-                                                  <Button
-                                                            className={'pull-right ml-1'}
-                                                            outline
-                                                            size="sm"
-                                                            onClick={() => this.handleDataEntryClick('manually')}
-                                                        >
-                                                            Edit
-                                                  </Button>
-                                                    </h3>
-                                                    <Card body>
-                                                        <Table className="mb-0">
-                                                            <tbody>
-                                                                <tr className="table-borderless">
-                                                                    <td>
-                                                                        <strong>Paper title:</strong> {this.state.paperTitle}
-                                                                    </td>
-                                                                </tr>
-                                                                <tr>
-                                                                    <td>
-                                                                        <strong>Authors:</strong>{' '}
-                                                                        {this.state.paperAuthors.map((author, index) => (
-                                                                            <span key={index}>
-                                                                                {this.state.paperAuthors.length > index + 1
-                                                                                    ? author.label + ', '
-                                                                                    : author.label}
-                                                                            </span>
-                                                                        ))}
-                                                                    </td>
-                                                                </tr>
-                                                                <tr>
-                                                                    <td>
-                                                                        <strong>Publication date:</strong>{' '}
-                                                                        {this.state.paperPublicationMonth
-                                                                            ? moment(this.state.paperPublicationMonth, 'M').format('MMMM')
-                                                                            : ''}{' '}
-                                                                        {this.state.paperPublicationYear}
-                                                                    </td>
-                                                                </tr>
-                                                            </tbody>
-                                                        </Table>
-                                                    </Card>
-                                                </Container>
-                                            ) : (
-                                                    ''
-                                                )}
-                                        </CSSTransitionGroup>
+                                            <TransitionGroup>
+                                                {this.state.showLookupTable ? (
+                                                    <Container
+                                                        key={1}
+                                                        classNames="slideDown"
+                                                        timeout={{ enter: 500, exit: 300 }}
+                                                    >
+                                                        <div className="mt-5">
+                                                            <h3 className="h4 mb-3">
+                                                                Lookup result
+                                                            <Button
+                                                                    className={'pull-right ml-1'}
+                                                                    outline
+                                                                    size="sm"
+                                                                    onClick={() => this.handleDataEntryClick('manually')}
+                                                            >
+                                                                    Edit
+                                                            </Button>
+                                                            </h3>
+                                                            <Card body>
+                                                                <Table className="mb-0">
+                                                                    <tbody>
+                                                                        <tr className="table-borderless">
+                                                                            <td>
+                                                                                <strong>Paper title:</strong> {this.state.paperTitle}
+                                                                            </td>
+                                                                        </tr>
+                                                                        <tr>
+                                                                            <td>
+                                                                                <strong>Authors:</strong>{' '}
+                                                                                {this.state.paperAuthors.map((author, index) => (
+                                                                                    <span key={index}>
+                                                                                        {this.state.paperAuthors.length > index + 1
+                                                                                            ? author.label + ', '
+                                                                                            : author.label}
+                                                                                    </span>
+                                                                                ))}
+                                                                            </td>
+                                                                        </tr>
+                                                                        <tr>
+                                                                            <td>
+                                                                                <strong>Publication date:</strong>{' '}
+                                                                                {this.state.paperPublicationMonth
+                                                                                    ? moment(this.state.paperPublicationMonth, 'M').format('MMMM')
+                                                                                    : ''}{' '}
+                                                                                {this.state.paperPublicationYear}
+                                                                            </td>
+                                                                        </tr>
+                                                                    </tbody>
+                                                                </Table>
+                                                            </Card>
+                                                        </div>
+                                                    </Container>
+                                                ) : (
+                                                        ''
+                                                    )}
+                                            </TransitionGroup>
+                                        </div>
                                     </Container>
                                 );
                             default:
                                 //Manually
                                 return (
-                                    <Container key={2}>
+                                    <Container key={2}
+                                        classNames="fadeIn"
+                                        timeout={{ enter: 500, exit: 0 }}
+                                    >
                                         <Form className="mt-4" onSubmit={this.submitHandler}>
                                             <FormGroup>
                                                 <Label for="paperTitle">
@@ -469,6 +500,9 @@ class GeneralData extends Component {
                                                                     value={this.state.paperPublicationMonth}
                                                                     onChange={this.handleMonthChange}
                                                                 >
+                                                                    <option value="" key="">
+                                                                        Month
+                                                                    </option>
                                                                     {moment.months().map((el, index) => {
                                                                         return (
                                                                             <option value={index + 1} key={index + 1}>
@@ -486,6 +520,9 @@ class GeneralData extends Component {
                                                                     value={this.state.paperPublicationYear}
                                                                     onChange={this.handleInputChange}
                                                                 >
+                                                                    <option value="" key="">
+                                                                        Year
+                                                                    </option>
                                                                     {range(1900, moment().year())
                                                                         .reverse()
                                                                         .map((year) => (
@@ -502,9 +539,15 @@ class GeneralData extends Component {
                                 );
                         }
                     })()}
-                </CSSTransitionGroup>
+                </TransitionGroup>
                 <hr className="mt-5 mb-3" />
-
+                {this.state.errors && this.state.errors.length > 0 &&
+                    <ul className="float-left mb-4 text-danger">
+                        {this.state.errors.map(e => {
+                            return <li>{e}</li>
+                        })}
+                    </ul>
+                }
                 <Button
                     color="primary"
                     className="float-right mb-4"
@@ -573,8 +616,14 @@ GeneralData.propTypes = {
     doi: PropTypes.string.isRequired,
     title: PropTypes.string.isRequired,
     authors: PropTypes.array.isRequired,
-    publicationMonth: PropTypes.number.isRequired,
-    publicationYear: PropTypes.number.isRequired,
+    publicationMonth: PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.number
+    ]).isRequired,
+    publicationYear: PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.number
+    ]).isRequired,
     showLookupTable: PropTypes.bool.isRequired,
     updateGeneralData: PropTypes.func.isRequired,
     nextStep: PropTypes.func.isRequired,
