@@ -1,9 +1,9 @@
+/* eslint-disable */
 import React, {Component} from 'react';
 import LinkButton from './LinkButton';
 import {Input} from 'reactstrap';
 import {updateResource} from '../network';
-import {NotificationManager} from 'react-notifications';
-import {popupDelay} from '../utils';
+import { toast } from 'react-toastify';
 
 export default class EditableHeader extends Component {
 
@@ -28,12 +28,12 @@ export default class EditableHeader extends Component {
         try {
             await updateResource(this.props.id, this.state.value);
             event.value = this.state.value;
-            NotificationManager.success('Resource name updated successfully', 'Success', popupDelay);
+            toast.success('Resource name updated successfully');
             this.props.onChange(event);
             this.setState({editorState: 'view'});
         } catch (error) {
             console.error(error);
-            NotificationManager.error(error.message, 'Error updating resource', popupDelay);
+            toast.error(`Error updating resource : ${error.message}`);
             this.setState({editorState: 'view'});
         }
     };
@@ -54,32 +54,36 @@ export default class EditableHeader extends Component {
                     <h1 key="h" className="h2">{this.state.value}</h1>,
                     <span key="span" className="toolbar toolbar-container toolbar-container-header">
                         <LinkButton value="edit" className="toolbar-button" spanClassName="fa fa-pencil"
-                                onClick={this.handleEditClick}/>
+                                onClick={this.handleEditClick}
+                        />
                     </span>,
                 ];
                 break;
             }
             case 'edit': {
-                content = <div className="snakView-header">
+                content = (<div className="snakView-header">
                     <div className="snakView-value snakView-variation-valueSnak">
                         <div className="valueView valueView-inEditMode">
                             <div className="valueView-value">
                                 <Input className="valueView-input-header valueView-input" value={this.state.value}
-                                        onChange={this.handleChange}/>
+                                        onChange={this.handleChange}
+                                />
                             </div>
                         </div>
                     </div>
                     <span className="toolbar toolbar-container toolbar-container-header">
                         <LinkButton value="publish" className="toolbar-container toolbar-button"
-                                spanClassName="fa fa-check" onClick={this.handleSubmitClick}/>
+                                spanClassName="fa fa-check" onClick={this.handleSubmitClick}
+                        />
                         <LinkButton value="cancel" className="toolbar-container toolbar-button"
-                                spanClassName="fa fa-close" onClick={this.handleCancelClick}/>
+                                spanClassName="fa fa-close" onClick={this.handleCancelClick}
+                        />
                     </span>
-                </div>;
+                           </div>);
                 break;
             }
             case 'loading': {
-                content = <span className="fa fa-spinner fa-spin"/>;
+                content = <span className="fa fa-spinner fa-spin" />;
                 break;
             }
             default: {
@@ -87,10 +91,11 @@ export default class EditableHeader extends Component {
             }
         }
 
-        return <div className="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center
-                        pb-2 mb-3 border-bottom">
+        return (<div className="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center
+                        pb-2 mb-3 border-bottom"
+                >
             {content}
-        </div>
+                </div>)
     }
 
 }
