@@ -13,27 +13,16 @@ class Home extends Component {
 
     render = () => {
         const showWarning = process.env.REACT_APP_SERVER_URL.includes('labs.tib.eu');
-        const showMouseflowWarning = process.env.REACT_APP_MOUSEFLOW_WEBSITE_ID ? true : false;
-
+        
         return (
             <div>
                 <Container className="p-0">
                     <h1 className="h4 mt-4 mb-4">Homepage</h1>
                 </Container>
                 <Container className="box pt-4 pb-4 pl-5 pr-5">
-                    {(showWarning || showMouseflowWarning) &&
+                    {showWarning &&
                         <Alert color="info">
-                            {showWarning &&
-                                <p className="p-0 m-0">
-                                    <strong>Warning: </strong> The ORKG is currently in an alpha stage. Data you enter in the system can be deleted without any notice.
-                                </p>
-                            }
-                            {showMouseflowWarning &&
-                                <p className="p-0 m-0">
-                                    <strong>Notice: </strong> The ORKG is using a third party analytics tool that is fully compliant with the GDPR. We collect this data anonymously for experimental purpose to enhance the user interface. If you do not wish to be tracked, follow this link:{' '}
-                                    <a target="_blank" rel="noopener noreferrer" href={'https://mouseflow.com/opt-out/'}>Opt Out <Icon size="sm" icon={faExternalLinkAlt} /></a>.
-                                </p>
-                            }
+                            <strong>Warning: </strong> The ORKG is currently in an alpha stage. Data you enter in the system can be deleted without any notice.
                         </Alert>
                     }
                     <Icon icon={faInfoCircle} className="text-primary" /> The <strong>Open Research Knowledge Graph</strong> - or - ORKG aims to describe research papers and contributions in a structured manner. With ORKG research contributions become findable and comparable. In order to add your own research, or to contribute,
