@@ -1,8 +1,10 @@
 import React, { Component } from 'react';
 import '../../assets/scss/DefaultLayout.scss';
-import { NotificationContainer } from 'react-notifications';
+import { ToastContainer, Slide } from 'react-toastify';
+import { popupDelay } from '../../utils';
 import Header from './Header/Header';
 import Footer from './Footer';
+import PropTypes from 'prop-types';
 
 class DefaultLayout extends Component {
     constructor(props) {
@@ -22,7 +24,12 @@ class DefaultLayout extends Component {
     render() {
         return (
             <div className="body">
-                <NotificationContainer />
+                <ToastContainer
+                    position="top-right"
+                    autoClose={parseInt(popupDelay)}
+                    hideProgressBar
+                    transition={Slide}
+                />
 
                 <Header />
                 {this.props.children}
@@ -31,5 +38,9 @@ class DefaultLayout extends Component {
         );
     }
 }
+
+DefaultLayout.propTypes = {
+    children: PropTypes.array.isRequired,
+};
 
 export default DefaultLayout;
