@@ -14,78 +14,105 @@ class ResearchField extends Component {
 
     constructor(props) {
         super(props);
+
+        this.pageSize = 25;
+
         this.state = {
             loading: true,
+            isNextPageLoading: false,
+            hasNextPage: false,
+            page: 1,
             researchField: null,
-            papers: null,
-            parentResearchField: null
+            papers: [],
+            parentResearchField: null,
+            isLastPageReached: false,
         };
     }
 
     componentDidMount() {
         this.loadResearchFieldData();
+        this.loadMorePapers();
     }
 
     componentDidUpdate = (prevProps) => {
         if (this.props.match.params.researchFieldId !== prevProps.match.params.researchFieldId) {
             this.loadResearchFieldData();
+            this.loadMorePapers();
         }
     }
 
-    loadResearchFieldData() {
-        // Get the research problem
+    loadResearchFieldData = () => {
+        // Get the research field
         getResource(this.props.match.params.researchFieldId).then((result) => {
-            this.setState({ researchField: result })
+            this.setState({ researchField: result, papers: [], loading: false })
             document.title = `${this.state.researchField.label} - ORKG`
         });
+    }
+
+    loadMorePapers = () => {
+        this.setState({ isNextPageLoading: true })
         // Get the statements that contains the research field as an object
         getStatementsByObject({
             id: this.props.match.params.researchFieldId,
-            limit: 4,
-            order: 'desc',
+            page: this.state.page,
+            items: this.pageSize,
+            sortBy: 'id',
+            desc: true
         }).then((result) => {
             // Papers
-            // Fetch the data of each paper
-            let papers = result.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_RESEARCH_FIELD)
-                .map((paper) => {
-                    return getStatementsBySubject(paper.subject.id).then((paperStatements) => {
-                        // publication year
-                        let publicationYear = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_YEAR);
-                        if (publicationYear.length > 0) {
-                            publicationYear = publicationYear[0].object.label
-                        }
-                        // publication month
-                        let publicationMonth = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_MONTH);
-                        if (publicationMonth.length > 0) {
-                            publicationMonth = publicationMonth[0].object.label
-                        } else {
-                            publicationMonth = ''
-                        }
-                        // authors
-                        let authors = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_AUTHOR);
-                        let authorNamesArray = [];
-                        if (authors.length > 0) {
-                            for (let author of authors) {
-                                let authorName = author.object.label;
-                                authorNamesArray.push(authorName);
+            if (result.length > 0) {
+                // Fetch the data of each paper
+                let papers = result.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_RESEARCH_FIELD)
+                    .map((paper) => {
+                        return getStatementsBySubject({ id: paper.subject.id }).then((paperStatements) => {
+                            // publication year
+                            let publicationYear = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_YEAR);
+                            if (publicationYear.length > 0) {
+                                publicationYear = publicationYear[0].object.label
+                            } else {
+                                publicationYear = ''
                             }
-                        }
-                        paper.data = {
-                            publicationYear,
-                            publicationMonth,
-                            authorNames: authorNamesArray.reverse(),
-                        }
-                        return paper;
+                            // publication month
+                            let publicationMonth = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_PUBLICATION_MONTH);
+                            if (publicationMonth.length > 0) {
+                                publicationMonth = publicationMonth[0].object.label
+                            } else {
+                                publicationMonth = ''
+                            }
+                            // authors
+                            let authors = paperStatements.filter((statement) => statement.predicate.id === process.env.REACT_APP_PREDICATES_HAS_AUTHOR);
+                            let authorNamesArray = [];
+                            if (authors.length > 0) {
+                                for (let author of authors) {
+                                    let authorName = author.object.label;
+                                    authorNamesArray.push(authorName);
+                                }
+                            }
+                            paper.data = {
+                                publicationYear,
+                                publicationMonth,
+                                authorNames: authorNamesArray.reverse(),
+                            }
+                            return paper;
+                        })
+                    });
+                let parentResearchField = result.find((statement) => statement.predicate.id === 'P36');
+                return Promise.all(papers).then((papers) => {
+                    this.setState({
+                        papers: [...this.state.papers, ...papers],
+                        parentResearchField: parentResearchField,
+                        isNextPageLoading: false,
+                        hasNextPage: (papers.length < this.pageSize || papers.length === 0) ? false : true,
+                        page: this.state.page + 1
                     })
-                });
-            let parentResearchField = result.find((statement) => statement.predicate.id === 'P36');
-            return Promise.all(papers).then((papers) => {
-                this.setState({
-                    papers: papers,
-                    loading: false,
-                    parentResearchField: parentResearchField
                 })
-            })
+            } else {
+                this.setState({
+                    isNextPageLoading: false,
+                    hasNextPage: false,
+                    isLastPageReached: true
+                });
+            }
         })
     }
 
@@ -102,7 +129,9 @@ class ResearchField extends Component {
                         <Container>
                             <Card>
                                 <CardHeader>
-                                    <div className="float-right"><b>{this.state.papers.length}</b> Papers</div>
+                                    {/* TODO: Show the total number of papers when number of items is provided with the paginated result
+                                        <div className="float-right"><b>{this.state.papers.length}</b> Papers</div>
+                                    */}
                                     <h3 className="h4 mt-4 mb-4">{this.state.researchField && this.state.researchField.label}</h3>
                                 </CardHeader>
                                 <CardBody>
@@ -123,7 +152,7 @@ class ResearchField extends Component {
                         </Container>
                         <br />
                         <Container>
-                            {this.state.papers && this.state.papers.length > 0 ?
+                            {this.state.papers.length > 0 &&
                                 <div>
                                     {this.state.papers.map(
                                         (resource) => {
@@ -136,7 +165,9 @@ class ResearchField extends Component {
                                         }
                                     )}
                                 </div>
-                                : (
+                            }
+                            {this.state.papers.length === 0 && !this.state.isNextPageLoading &&
+                                (
                                     <div className="text-center mt-4 mb-4">
                                         There are no articles for this research field, yet.
                                         <br />
@@ -148,6 +179,17 @@ class ResearchField extends Component {
                                         </Link>
                                     </div>
                                 )
+                            }
+                            {this.state.isNextPageLoading && <div className="text-center mt-4 mb-4"><Icon icon={faSpinner} spin /> Loading</div>}
+                            {!this.state.isNextPageLoading && this.state.hasNextPage && (
+                                <div style={{ cursor: 'pointer' }} className="list-group-item list-group-item-action text-center mt-2" onClick={!this.state.isNextPageLoading ? this.loadMorePapers : undefined}>
+                                    Load more papers
+                                </div>
+                            )}
+                            {!this.state.hasNextPage && this.state.isLastPageReached && (
+                                <div className="text-center mt-3">
+                                    You have reached the last page.
+                                </div>)
                             }
                         </Container>
                     </div>
