@@ -4,7 +4,7 @@ import { StyledContribution, StyledContributionsList } from '../AddPaper/Contrib
 import { getResource, getSimilaireContribution } from '../../network';
 
 import AddToComparison from './AddToComparison';
-import { CSSTransitionGroup } from 'react-transition-group'
+import { TransitionGroup, CSSTransition } from 'react-transition-group';
 import ContentLoader from 'react-content-loader'
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
@@ -30,7 +30,7 @@ const Title = styled.div`
     }
 `;
 
-const AnimationContainer = styled.div`
+const AnimationContainer = styled(CSSTransition)`
     transition: 0.3s background-color,  0.3s border-color;
 
     &.fadeIn-enter {
@@ -127,7 +127,7 @@ class Contributions extends Component {
                                     {this.props.contributions.map((contribution, index) => {
                                         return (
                                             <li className={contribution.id === selectedContributionId ? 'activeContribution' : ''} key={contribution.id}>
-                                                <Link to={reverse(ROUTES.VIEW_PAPER_CONTRIBUTION, { resourceId : this.props.paperId, contributionId : contribution.id})} className={'selectContribution'}>
+                                                <Link to={reverse(ROUTES.VIEW_PAPER_CONTRIBUTION, { resourceId: this.props.paperId, contributionId: contribution.id })} className={'selectContribution'}>
                                                     {contribution.label}
                                                 </Link>
                                             </li>
@@ -136,15 +136,14 @@ class Contributions extends Component {
                                 </StyledContributionsList>
                             )}
                         </Col>
-                        <CSSTransitionGroup
-                            transitionName="fadeIn"
-                            transitionEnterTimeout={500}
-                            transitionLeave={false}
-                            component="div"
+                        <TransitionGroup
                             className="col-9"
+                            exit={false}
                         >
                             <AnimationContainer
                                 key={selectedContributionId}
+                                classNames="fadeIn"
+                                timeout={{ enter: 500, exit: 0 }}
                             >
                                 <StyledContribution>
                                     {!this.state.loading && (
@@ -176,8 +175,8 @@ class Contributions extends Component {
                                                 <>
                                                     {this.props.researchProblems[selectedContributionId] && this.props.researchProblems[selectedContributionId].map((problem, index) => (
                                                         <span key={index}>
-                                                            <Link to={reverse(ROUTES.RESEARCH_PROBLEM, {researchProblemId : problem.id})}>
-                                                                <span className="btn btn-link p-0 border-0 align-baseline">
+                                                            <Link to={reverse(ROUTES.RESEARCH_PROBLEM, { researchProblemId: problem.id })}>
+                                                                <span style={{ whiteSpace: 'normal', textAlign: 'left' }} className="btn btn-link p-0 border-0 align-baseline">
                                                                     {problem.label}
                                                                 </span>
                                                             </Link>
@@ -249,7 +248,7 @@ class Contributions extends Component {
                                 </StyledContribution>
 
                             </AnimationContainer>
-                        </CSSTransitionGroup>
+                        </TransitionGroup>
                     </Row>
                 </Container>
             </div>
