@@ -7,7 +7,7 @@ import { Modal, ModalHeader, ModalBody, Input, Form, FormGroup, Label } from 're
 import uniqBy from 'lodash/uniqBy';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
-import 'vis/dist/vis-network.min.css';
+import 'vis-network/dist/vis-network.min.css';
 
 class GraphView extends Component {
     state = {
@@ -158,7 +158,7 @@ class GraphView extends Component {
             return list;
         }
 
-        let statements = await getStatementsBySubject(resourceId);
+        let statements = await getStatementsBySubject({ id: resourceId });
 
         if (statements.length > 0) {
             list.push(...statements);
@@ -234,7 +234,7 @@ class GraphView extends Component {
                         <Graph
                             graph={graph}
                             options={options}
-                            //events={events}
+                        //events={events}
                         />
                     )}
                     {this.state.isLoadingStatements && (
