@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 import React, {Component, Fragment} from 'react';
 import ObjectTypeSelector from './ObjectTypeSelector';
 import EditableDropdown from './new/EditableDropdown';
@@ -38,7 +40,7 @@ export default class MainSnak extends Component {
 
     getInput(isLiteral, inputStyle) {
         if (isLiteral) {
-            return <Fragment>
+            return (<Fragment>
                 <div className="valueView-input-group-prepend">
                     <span className="valueView-input-group-text">&quot;</span>
                 </div>
@@ -49,81 +51,83 @@ export default class MainSnak extends Component {
                                 <textarea className="valueView-input" defaultValue={this.props.text}
                                         style={inputStyle} onInput={this.props.onInput}
                                         autoFocus={true}
-                                        onKeyUp={this.props.onTextAreaKeyUp}>
-                                </textarea>
+                                        onKeyUp={this.props.onTextAreaKeyUp}
+                                />
                             </div>
                         </div>
                     </div>
-                    <div className="snakView-indicators"/>
+                    <div className="snakView-indicators" />
                 </div>
                 <div className="valueView-input-group-prepend">
                     <span className="valueView-input-group-text">&quot;</span>
                 </div>
-            </Fragment>;
+                    </Fragment>);
         } else {
-            return <div className="snakView-body">
+            return (<div className="snakView-body">
                 <div className="snakView-value snakView-variation-valueSnak ">
                     <div className="valueView valueView-inEditMode">
                         <div className="valueView-value">
                             <EditableDropdown requestUrl={resourcesUrl} placeholder="object" value={this.props.text}
                                     onItemSelected={this.handleObjectSelect} onInput={this.props.onInput}
-                                    onKeyUp={this.props.onKeyUp}/>
+                                    onKeyUp={this.props.onKeyUp}
+                            />
                         </div>
                     </div>
                 </div>
-                <div className="snakView-indicators"/>
-            </div>;
+                <div className="snakView-indicators" />
+                    </div>);
         }
     }
 
     render() {
         let content = null;
         if (!this.props.editing) {
-            content = <div className="snakView">
+            content = (<div className="snakView">
                 <div className="snakView-value-container">
-                    <div className="snakView-typeSelector"/>
+                    <div className="snakView-typeSelector" />
                     <div className="snakView-body">
                         <div className="snakView-value">
                             {
                                 this.props.objectType === 'resource'
                                 ? <a href={`${process.env.PUBLIC_URL}/resource/${encodeURIComponent(this.props.id)}`}>
                                     {this.props.text}
-                                </a>
+                                  </a>
                                 : this.props.text
                             }
                         </div>
-                        <div className="snakView-indicators"/>
+                        <div className="snakView-indicators" />
                     </div>
                 </div>
-            </div>
+                       </div>)
         } else {
-            const inputStyle = {height: "21.8px", overflow: "hidden", resize: "none"};
+            const inputStyle = {height: '21.8px', overflow: 'hidden', resize: 'none'};
             const shouldShowQuotes = this.props.objectType === 'literal';
-            content = <div className="snakView edit" aria-disabled="false">
+            content = (<div className="snakView edit" aria-disabled="false">
                 {
                     this.props.newProperty && <div className="snakView-property-container">
                         <div className="snakView-property" dir="auto">
                             <EditableDropdown requestUrl={predicatesUrl} placeholder="property"
                                     onItemSelected={this.handlePropertySelect}
                                     onNewItemSelected={this.handleNewProperty}
-                                    onKeyUp={() => {}}/>
+                                    onKeyUp={() => {}}
+                            />
                         </div>
-                    </div>
+                                              </div>
                 }
                 {
                     !this.props.newProperty || this.state.selectedPredicateId || this.state.newPredicateLabel ?
                     <div className="snakView-value-container" dir="auto">
-                        <ObjectTypeSelector onItemSelect={this.handleItemSelect} objectType={this.props.objectType}/>
+                        <ObjectTypeSelector onItemSelect={this.handleItemSelect} objectType={this.props.objectType} />
                         {this.getInput(shouldShowQuotes, inputStyle)}
                     </div>
                     : null
                 }
-            </div>
+                       </div>)
         }
 
-        return <div className="statementView-mainSnak">
+        return (<div className="statementView-mainSnak">
             {content}
-        </div>
+                </div>)
     }
 
 }
