@@ -19,6 +19,7 @@ export default {
    /* Legacy routes, only used for debugging now */
    RESSOURCE_DETAILS: '/resource/:resourceId/:sectionName',
    PREDICATE_DETAILS: '/predicate/:predicateId',
-   SEARCH: '/search/:searchTerm',
+   SEARCH: '/search/:searchTerm?',
    TPDL: '/tpdl',
+   STATS: '/stats',
 }
