@@ -1,3 +1,4 @@
+import queryString from 'query-string';
 export const url = process.env.REACT_APP_SERVER_URL;
 export const similaireServiceUrl = process.env.REACT_APP_SIMILARITY_SERVICE_URL;
 export const annotationServiceUrl = process.env.REACT_APP_ANNOTATION_SERVICE_URL;
@@ -6,6 +7,7 @@ export const predicatesUrl = `${url}predicates/`;
 export const statementsUrl = `${url}statements/`;
 export const literalsUrl = `${url}literals/`;
 export const classesUrl = `${url}classes/`;
+export const statsUrl = `${url}stats/`;
 export const crossrefUrl = process.env.REACT_APP_CROSSREF_URL;
 export const arxivUrl = process.env.REACT_APP_ARXIV_URL;
 export const semanticScholarUrl = process.env.REACT_APP_SEMANTICSCHOLAR_URL;
@@ -104,8 +106,8 @@ export const updateLiteral = (id, label) => {
   );
 };
 
-export const createResource = (label) => {
-  return submitPostRequest(resourcesUrl, { 'Content-Type': 'application/json' }, { label: label });
+export const createResource = (label, classes = []) => {
+  return submitPostRequest(resourcesUrl, { 'Content-Type': 'application/json' }, { label, classes });
 };
 
 export const createLiteral = (label) => {
@@ -154,43 +156,72 @@ export const getResource = (id) => {
   return submitGetRequest(`${resourcesUrl}${encodeURIComponent(id)}/`);
 };
 
-export const getAllResources = () => {
-  return submitGetRequest(resourcesUrl);
+export const getAllResources = ({ page = 1, items = 9999, sortBy = 'id', desc = true }) => {
+
+  let params = queryString.stringify({ page: page, items: items, sortBy: sortBy, desc: desc })
+
+  return submitGetRequest(`${resourcesUrl}?${params}`);
 };
 
-export const getAllStatements = () => {
-  return submitGetRequest(statementsUrl);
+export const getAllStatements = ({ page = 1, items = 9999, sortBy = 'id', desc = true }) => {
+
+  let params = queryString.stringify({ page: page, items: items, sortBy: sortBy, desc: desc })
+
+  return submitGetRequest(`${statementsUrl}?${params}`);
 };
 
 export const getPredicatesByLabel = (label) => {
   return submitGetRequest(predicatesUrl + '?q=' + encodeURIComponent(label));
 };
 
-export const getStatementsBySubject = (id) => {
-  return submitGetRequest(`${statementsUrl}subject/${encodeURIComponent(id)}/`);
+export const getStatementsBySubject = ({ id, page = 1, items = 9999, sortBy = 'id', desc = true }) => {
+
+  let params = queryString.stringify({ page: page, items: items, sortBy: sortBy, desc: desc })
+
+  return submitGetRequest(`${statementsUrl}subject/${encodeURIComponent(id)}/?${params}`);
 };
 
-export const getStatementsByObject = async ({ id, order = 'asc', limit = null }) => {
-  let statements = await submitGetRequest(`${statementsUrl}object/${encodeURIComponent(id)}/`);
+export const getStatementsByObject = async ({ id, page = 1, items = 9999, sortBy = 'id', desc = true }) => {
+
+  let params = queryString.stringify({ page: page, items: items, sortBy: sortBy, desc: desc })
+
+  let statements = await submitGetRequest(`${statementsUrl}object/${encodeURIComponent(id)}/?${params}`);
 
   // TODO: replace sorting and limit by backend functionalities when ready
   statements.sort((a, b) => {
-    if (order === 'asc') {
+    if (!desc) {
       return parseInt(a.id.replace('S', '')) - parseInt(b.id.replace('S', ''));
     } else {
       return parseInt(b.id.replace('S', '')) - parseInt(a.id.replace('S', ''));
     }
   });
 
-  if (limit) {
-    statements = statements.slice(0, limit);
-  }
-
   return statements;
 };
 
-export const getStatementsByPredicate = (id) => {
-  return submitGetRequest(`${statementsUrl}predicate/${encodeURIComponent(id)}/`);
+export const getResourcesByClass = async ({ id, page = 1, items = 9999, sortBy = 'id', desc = true }) => {
+
+  let params = queryString.stringify({ page: page, items: items, sortBy: sortBy, desc: desc })
+
+  let resources = await submitGetRequest(`${classesUrl}${encodeURIComponent(id)}/resources/?${params}`);
+
+  // TODO: replace sorting and limit by backend functionalities when ready
+  resources.sort((a, b) => {
+    if (!desc) {
+      return parseInt(a.id.replace('R', '')) - parseInt(b.id.replace('R', ''));
+    } else {
+      return parseInt(b.id.replace('R', '')) - parseInt(a.id.replace('R', ''));
+    }
+  });
+
+  return resources;
+};
+
+export const getStatementsByPredicate = ({ id, page = 1, items = 9999, sortBy = 'id', desc = true }) => {
+
+  let params = queryString.stringify({ page: page, items: items, sortBy: sortBy, desc: desc })
+
+  return submitGetRequest(`${statementsUrl}predicate/${encodeURIComponent(id)}/?${params}`);
 };
 
 export const getSimilaireContribution = (id) => {
@@ -198,15 +229,17 @@ export const getSimilaireContribution = (id) => {
 };
 
 export const getAnnotations = (abstract) => {
-  return submitGetRequest(
-    `${annotationServiceUrl}annotator/?text2annotate=${encodeURIComponent(abstract)}`,
-  );
+  return submitPostRequest(`${annotationServiceUrl}annotator/`, { 'Content-Type': 'application/json' }, { text2annotate: abstract });
 };
 
 export const indexContribution = (contribution_id) => {
   return fetch(`${similaireServiceUrl}internal/index/${encodeURIComponent(contribution_id)}/`, {
     method: 'GET',
   });
+};
+
+export const getStats = () => {
+  return submitGetRequest(statsUrl);
 };
 
 export const createShortLink = (data) => {
@@ -223,4 +256,8 @@ export const getLongLink = (shortCode) => {
 
 export const getAllClasses = () => {
   return submitGetRequest(classesUrl);
+};
+
+export const saveFullPaper = (data) => {
+  return submitPostRequest(`${url}papers/`, { 'Content-Type': 'application/json' }, data);
 };
