@@ -6,13 +6,14 @@ import { Cookies } from 'react-cookie';
 const initialState = {
     isTourOpen: false,
     currentStep: 1,
+    shouldBlockNavigation: false,
     tourCurrentStep: 1,
     tourStartAt: 0,
     title: '',
     authors: [],
     abstract: '',
-    publicationMonth: 1,
-    publicationYear: 2000,
+    publicationMonth: '',
+    publicationYear: '',
     entry: '',
     showLookupTable: false,
     doi: '',
@@ -55,6 +56,20 @@ export default (state = initialState, action) => {
             return {
                 ...state,
                 currentStep: state.currentStep - 1,
+            };
+        }
+
+        case type.ADD_PAPER_BLOCK_NAVIGATION: {
+            return {
+                ...state,
+                shouldBlockNavigation: true,
+            };
+        }
+
+        case type.ADD_PAPER_UNBLOCK_NAVIGATION: {
+            return {
+                ...state,
+                shouldBlockNavigation: false,
             };
         }
 
