@@ -2,7 +2,7 @@ export const doiPredicateLabel = process.env.REACT_APP_DOI_PREDICATE_LABEL;
 export const popupDelay = process.env.REACT_APP_POPUP_DELAY;
 
 export function hashCode(s) {
-    return s.split("").reduce((a, b) => {
+    return s.split('').reduce((a, b) => {
         a = ((a << 5) - a) + b.charCodeAt(0);
         return a & a;
     }, 0);
